step_1/jobindex_paginator/requirements.txt
azure-core==1.30.2
azure-cosmos==4.7.0
azure-functions==1.20.0
azure-servicebus
azure-storage-blob==12.20.0 
cffi==1.16.0 
cryptography==42.0.8 
isodate==0.6.1 
pycparser==2.22
beautifulsoup4==4.12.3
common @ git+https://github.com/it-minds/itm_leads-finder-scripts.git@main#subdirectory=internal
requests==2.32.3
six==1.16.0
soupsieve==2.5
typing_extensions==4.12.1
urllib3==2.2.1